Made improvements to fix bug ID3072 on cbuffUngetByte function, and bug ID3498 to test function.
Fixed the Test 20a to correctly test the cbuffUngetByte function. Current reason for test failure is that there is still old data in the buffer. The Unget function will continue to return old data to the buffer *until* the tail pointer wraps back to the head pointer. This is how this function should work.
Basically, cbuffUngetByte does, in principle, actually work!
Edited the cbuff.c, function cbuffUngetByte function to remove unnecessary test for and EMPTY buffer. If buffer is empty, through unget will intend to change that state by returning the most recent value back to the buffer. Thus this test is redundant. Currently only #if0/#endif'ed out. Can be deleted with next commit. Also added code to set the "buffer not empty" flag, as after this function completes, the buffe is no longer empty.
Also added to the function header describing the case that the buffer could read out old, random data if the buffer has not been completely filled at some point.
Fixed typo in the comments.



Signed-off-by: Stuart Cording <codinghead@gmail.com>