Fixed bug ID3073 and bug ID3527.
Completed modifications to cbuff-module to speed up the creation and desctruction of buffer objects.
cbuffCreate and cbuffDestroy are now completely changed.
Bug ID3073 is now fixed.

Added a new test, test 1a, that ensures that none of the 16 buffers can be destroyed if they were not created.
Bug ID3527 is now fixed.

Signed-off-by: Stuart Cording <codinghead@gmail.com>