Attempt to fix cbuff latch-up.
Moved point at which the 'buffer empty' flag is cleared in cbuffPutByte
in an effort to avoid latch-up seen on interrupt based usage of this
function. This line of code occurs at same point as in cbuffGetByte which
was previously not the case.

This fix may or may not work - it is only a trial.